Delete old RangeMap; move CAMap to shared.
